better mushroomfx, character blend script notes
I need to remove CoreIK, or find a way to get it to import. I am not sure where the import chokes, but the entire setup feels somewhat unorthodox. unfortunately, some of my animations are using CoreIK now and I might want to redo them without it. Not having it complicates some matters (particularly, inherited rotation for child bones down the chain), I need to find a cleaner solution for working with body twist etc